Factor RemoteSplit into fs
